fix: move test file location, use langchain google gen ai, update docstring and update jinja template file name to include promting technique used
